+++ frontend-next-migration/src/entities/Hero/ui/HeroContainerV3/HeroContainer.tsx
